Update lichess_study_2023-world-championship-nepomniachtchi-ding.pgn
I tried to run the analysis.py with its pgn and gave a series of errors, so I went to check and it was that the annotation was badly diagrammed. 

So I did the following, accessed the site of lichess (https://lichess.org/api/broadcast/lrdHUzyS.pgn) and downloaded the pgn straight from there, deleted the matches that will not happen yet and ran the analysis.py and everything worked out.